Added OS X support
