+++ Source/UIView+JRTAutoLayout.h
@@ -1,0 +1,13 @@
+//
+//  NSView+JRTAutoLayout.h
+//
+//  Created by Vanderlei Martinelli on 31/05/2014.
+//
+//  Convenience header file.
+//  You can use "JRTView+JRTAutoLayout.h" if you want.
+//
+
+#if TARGET_OS_IPHONE
+#import "JRTView+JRTAutoLayout.h"
+#endif
+
